Package for PyPi
This commit allows the module to be packaged for PyPi. Additionally, as
there is currently a clash with the name "Matter", the module has been
renamed to "Airer".

Tests are no longer under the module, and now reside outside of it.
Tests are now intended to be run with `setup.py test`, which invokes
`pytest` under the hood.